Vasts changes and fixes. Starting to sort through code.
