+++ pyproject.toml
@@ -1,6 +1,14 @@
-
-[project]
+[tool.poetry]
 name = "odorsampling"
-authors = [
-    {name = "Christopher De Jesus", email = "cd525@cornell.edu"}
-]
+version = "0.1.0"
+description = ""
+authors = ["Christopher De Jesus <Christopher2653@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.11"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
